maincore/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django== 4.2.13
django-allauth==0.63.3
django-filter==24.2
djangorestframework==3.15.1
drf-writable-nested==0.7.0
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.1
pip==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==49.2.1
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
xlrd==1.2.0
